# contracts/generated/openapi/contracts.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RiskItem ModelOps API",
    "version": "0.1.0",
    "description": "信贷风控模型智能监测与自主迭代系统 API。由 scripts/generate_openapi.py 从 FastAPI app 自动生成，不手工维护。"
  },
  "paths": {
...
    "/api/diagnosis/events/{event_id}/agent-handoff": {
      "get": {
        "tags": [
          "diagnosis"
        ],
        "summary": "Get Agent Handoff",
        "description": "Read-only Agent handoff contract. Delegates to DiagnosisHandoffService.",
        "operationId": "get_agent_handoff_api_diagnosis_events__event_id__agent_handoff_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnosis/runs/{diagnosis_run_id}": {
      "get": {
        "tags": [
          "diagnosis"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_diagnosis_runs__diagnosis_run_id__get",
        "parameters": [
          {
            "name": "diagnosis_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagnosis Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnosis/runs/by-monitoring/{monitoring_run_id}": {
      "get": {
        "tags": [
          "diagnosis"
        ],
        "summary": "Get Diagnosis By Monitoring",
        "description": "查询某个监控运行对应的最新诊断结果（含候选排序 + 证据详情）。",
        "operationId": "get_diagnosis_by_monitoring_api_diagnosis_runs_by_monitoring__monitoring_run_id__get",
        "parameters": [
          {
            "name": "monitoring_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monitoring Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnosis/trigger": {
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "Trigger Diagnosis",
        "description": "手动触发一次诊断（用于场景注入后的根因分析）。",
        "operationId": "trigger_diagnosis_api_diagnosis_trigger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerDiagnosisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TriggerDiagnosisRequest": {
        "properties": {
          "monitoring_run_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Monitoring Run Id"
          },
          "lifecycle_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Run Id"
          }
        },
        "type": "object",
        "required": [
          "monitoring_run_id"
        ],
        "title": "TriggerDiagnosisRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}